Github Workflow: Build and run tests for each PR to main
## Task 
To improve code stability I want the build and tests to run for each PR to main branch. 

## Solution
Include a new workflow yaml file that should instruct github actions to build and run all tests. 

## Testing
The proof is in the pudding - if it works we should see the correct behavior on this PR merge.