#14: Update Task Status (#23)
## Task 
In this PR we enable user to change the TaskStatus via a drop down menu on the Task card. 

## Solution
Implement a dropdown menu onClick of the Task's Status icon button. Also in this PR - modified the UI so that the Status and Delete icons are larger and in line with each other.

Moved the state for openness of dialog and menu to Presenters so they survive rotation.

## Testing
Auto build and unit tests pass. Manual testing of app functionality works